+++ CMakeLists.txt
@@ -1,19 +1,20 @@
 cmake_minimum_required(VERSION 3.10)
 project(netsock C CXX)
 
+set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+
 set(CMAKE_C_STANDARD 11)
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Werror -Wall")
 
 set(SRC
-    src/arguments/arguments.c
-    src/logger/logger.c
-    src/network/network.c
-    src/network/send/send.c
-    src/network/recv/recv.c
+    src/arguments.c
+    src/logger.c
+    src/network.c
+    src/send.c
+    src/recv.c
 )
 
 add_library(${PROJECT_NAME}_lib STATIC ${SRC})
-target_include_directories(${PROJECT_NAME}_lib PUBLIC src)
+target_include_directories(${PROJECT_NAME}_lib PUBLIC include)
 
 add_executable(${PROJECT_NAME} src/main.c)
 target_link_libraries(${PROJECT_NAME} PRIVATE ${PROJECT_NAME}_lib)
