+++ include/protocol.h
@@ -1,0 +1,8 @@
+#ifndef NETSOCK_PROTOCOL_H
+#define NETSOCK_PROTOCOL_H
+
+enum Protocol { TCP = 0, UDP = 1 };
+
+typedef enum Protocol protocol_t;
+
+#endif  // NETSOCK_PROTOCOL_H
