--- src/protocol/protocol.h
@@ -1,3 +1,0 @@
-enum Protocol { TCP = 0, UDP = 1 };
-
-typedef enum Protocol protocol_t;
