+++ src/recv.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <unistd.h>
 
-#include "../../logger/logger.h"
+#include "logger.h"
 
 typedef ssize_t (*recv_fun)(int fd, void *buffer, size_t n);
