+++ src/send.c
@@ -3,16 +3,16 @@
 #include <stdio.h>
 #include <unistd.h>
 
-#include "../../logger/logger.h"
+#include "logger.h"
 
-typedef ssize_t (*send_fun)(int fd, const void *buffer, size_t n,
-                            const sockaddr_t *dest);
+typedef ssize_t (*send_fun)(int fd, const void* buffer, size_t n,
+                            const sockaddr_t* dest);
 
-ssize_t udp_send(int fd, const void *buffer, size_t n, const sockaddr_t *dest);
+ssize_t udp_send(int fd, const void* buffer, size_t n, const sockaddr_t* dest);
 
-ssize_t tcp_send(int fd, const void *buffer, size_t n, const sockaddr_t *dest);
+ssize_t tcp_send(int fd, const void* buffer, size_t n, const sockaddr_t* dest);
 
-size_t send_stdin(int socket, const sockaddr_t *dest, send_fun send,
+size_t send_stdin(int socket, const sockaddr_t* dest, send_fun send,
                   size_t buf_size) {
   uint8_t buffer[buf_size];
 
@@ -35,7 +35,7 @@ size_t send_stdin(int socket, const sockaddr_t *dest, send_fun send,
   return sum_sent;
 }
 
-size_t send_in(int socket, const sockaddr_t *dest, const protocol_t *protocol,
+size_t send_in(int socket, const sockaddr_t* dest, const protocol_t* protocol,
                size_t buf_size) {
   send_fun send;
   switch (*protocol) {
@@ -49,10 +49,10 @@ size_t send_in(int socket, const sockaddr_t *dest, const protocol_t *protocol,
   return send_stdin(socket, dest, send, buf_size);
 }
 
-ssize_t udp_send(int fd, const void *buffer, size_t n, const sockaddr_t *dest) {
+ssize_t udp_send(int fd, const void* buffer, size_t n, const sockaddr_t* dest) {
   return sendto(fd, buffer, n, 0, &dest->addr, sizeof(sockaddr_t));
 }
 
-ssize_t tcp_send(int fd, const void *buffer, size_t n, const sockaddr_t *dest) {
+ssize_t tcp_send(int fd, const void* buffer, size_t n, const sockaddr_t* dest) {
   return send(fd, buffer, n, 0);
 }
