+++ test/test_arguments.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
 extern "C" {
-#include "arguments/arguments.h"
+#include "arguments.h"
 }
 
 TEST(ProjectTest, SomeFunction) { EXPECT_EQ(17, 17); }
